Disable test on windows
